+++ weave-js/src/common/css/color.styles.ts
@@ -1,0 +1,185 @@
+export const MOON = {
+  MOON_0: '#FFFFFF', // white
+  MOON_50: '#FAFAFA',
+  MOON_100: '#F5F5F5',
+  MOON_150: '#F0F0F0',
+  MOON_200: '#E6E6E6',
+  MOON_250: '#DDDDDD',
+  MOON_300: '#D1D1D1',
+  MOON_350: '#C2C2C4',
+  MOON_400: '#AFB0B2',
+  MOON_450: '#949699',
+  MOON_500: '#84868A',
+  MOON_550: '#74777A',
+  MOON_600: '#5D6063',
+  MOON_650: '#4D5154',
+  MOON_700: '#43474A',
+  MOON_750: '#3B3E42',
+  MOON_800: '#303438',
+  MOON_850: '#212529',
+  MOON_900: '#1D1F24',
+  MOON_950: '#171A1F',
+  MOON_1000: '#0E1014', // oblivion
+};
+
+export const NEW_TEAL = {
+  TEAL_200: '#E1F7FA',
+  TEAL_250: '#CDF4F7',
+  TEAL_300: '#A9EDF2',
+  TEAL_350: '#83E4EB',
+  TEAL_400: '#58D3DB',
+  TEAL_450: '#10BFCC',
+  TEAL_500: '#13A9BA',
+  TEAL_550: '#0097AB',
+  TEAL_600: '#038194',
+  TEAL_650: '#096A7A',
+  TEAL_700: '#155B69',
+  TEAL_750: '#18464F',
+  TEAL_800: '#1B3B42',
+};
+
+export const NEW_GREEN = {
+  GREEN_200: '#E4F7EE',
+  GREEN_250: '#C9F5E0',
+  GREEN_300: '#A1F0CB',
+  GREEN_350: '#85E5BC',
+  GREEN_400: '#5ED6A4',
+  GREEN_450: '#28C787',
+  GREEN_500: '#00B26E',
+  GREEN_550: '#009962',
+  GREEN_600: '#00875A',
+  GREEN_650: '#0B7052',
+  GREEN_700: '#165745',
+  GREEN_750: '#19473B',
+  GREEN_800: '#193831',
+};
+
+export const NEW_CACTUS = {
+  CACTUS_200: '#EFF7DF',
+  CACTUS_250: '#E0F5BA',
+  CACTUS_300: '#D0ED9D',
+  CACTUS_350: '#C6EB83',
+  CACTUS_400: '#BBE06B',
+  CACTUS_450: '#A5D154',
+  CACTUS_500: '#92C23A',
+  CACTUS_550: '#7AA638',
+  CACTUS_600: '#5F8A2D',
+  CACTUS_650: '#496B22',
+  CACTUS_700: '#3D541D',
+  CACTUS_750: '#34451C',
+  CACTUS_800: '#2B3819',
+};
+
+export const NEW_GOLD = {
+  GOLD_200: '#FEF1D8',
+  GOLD_250: '#FEECC1',
+  GOLD_300: '#FFE49E',
+  GOLD_350: '#FFDD80',
+  GOLD_400: '#FFCF4D',
+  GOLD_450: '#FAC13C',
+  GOLD_500: '#F5A822',
+  GOLD_550: '#D4870D',
+  GOLD_600: '#B8740F',
+  GOLD_650: '#8F5E14',
+  GOLD_700: '#694919',
+  GOLD_750: '#4D391C',
+  GOLD_800: '#3C2F1A',
+};
+
+export const NEW_SIENNA = {
+  SIENNA_200: '#FFEDE0',
+  SIENNA_250: '#FFE0CC',
+  SIENNA_300: '#FFCFB2',
+  SIENNA_350: '#FFBA91',
+  SIENNA_400: '#FCA36F',
+  SIENNA_450: '#FC8F58',
+  SIENNA_500: '#F27641',
+  SIENNA_550: '#D96534',
+  SIENNA_600: '#C2562F',
+  SIENNA_650: '#994228',
+  SIENNA_700: '#703726',
+  SIENNA_750: '#523026',
+  SIENNA_800: '#452B21',
+};
+
+export const NEW_RED = {
+  RED_200: '#FFE6E6',
+  RED_250: '#FFDBDC',
+  RED_300: '#FFC7CA',
+  RED_350: '#FFA8B0',
+  RED_400: '#FF7A88',
+  RED_450: '#FF596D',
+  RED_500: '#FF3D5A',
+  RED_550: '#E3324F',
+  RED_600: '#CC2944',
+  RED_650: '#A32740',
+  RED_700: '#782637',
+  RED_750: '#52212B',
+  RED_800: '#47222A',
+};
+
+export const NEW_MAGENTA = {
+  MAGENTA_200: '#F7E5FD',
+  MAGENTA_250: '#F3D5FD',
+  MAGENTA_300: '#EFC2FC',
+  MAGENTA_350: '#E7A7FC',
+  MAGENTA_400: '#E180FF',
+  MAGENTA_450: '#D869FA',
+  MAGENTA_500: '#CB51F0',
+  MAGENTA_550: '#BC45E5',
+  MAGENTA_600: '#9E36C2',
+  MAGENTA_650: '#7A2C94',
+  MAGENTA_700: '#602973',
+  MAGENTA_750: '#522761',
+  MAGENTA_800: '#42224D',
+};
+
+export const NEW_PURPLE = {
+  PURPLE_200: '#EDE8FF',
+  PURPLE_250: '#E4DBFF',
+  PURPLE_300: '#D6C9FF',
+  PURPLE_350: '#C2B1FC',
+  PURPLE_400: '#B199FF',
+  PURPLE_450: '#9D80FF',
+  PURPLE_500: '#8465F0',
+  PURPLE_550: '#7251E0',
+  PURPLE_600: '#6645D1',
+  PURPLE_650: '#5539B2',
+  PURPLE_700: '#4B3791',
+  PURPLE_750: '#3C2F6B',
+  PURPLE_800: '#352B59',
+};
+
+export const NEW_BLUE = {
+  BLUE_200: '#E0EDFE',
+  BLUE_250: '#D0E4FE',
+  BLUE_300: '#BDD9FF',
+  BLUE_350: '#A4C9FC',
+  BLUE_400: '#7DB1FA',
+  BLUE_450: '#629DF5',
+  BLUE_500: '#397EED',
+  BLUE_550: '#286CE0',
+  BLUE_600: '#1F59C4',
+  BLUE_650: '#19469E',
+  BLUE_700: '#193C80',
+  BLUE_750: '#1E376B',
+  BLUE_800: '#1D3157',
+};
+
+export const NEW_SLATE = {
+  SLATE_200: '#EFF3F9',
+  SLATE_250: '#E1E8F2',
+  SLATE_300: '#D5DFEB',
+  SLATE_350: '#C5CFDE',
+  SLATE_400: '#B8C1D1',
+  SLATE_450: '#A3ADC2',
+  SLATE_500: '#818EA3',
+  SLATE_550: '#677389',
+  SLATE_600: '#596375',
+  SLATE_650: '#464E5C',
+  SLATE_700: '#3A404D',
+  SLATE_750: '#323742',
+  SLATE_800: '#2B2E38',
+};
+
+export const COLORS = {};
